finished admin backend

+++ frontend/src/Components/AdminPage/ProductItem.js
@@ -26,7 +26,7 @@ const ProductItem = ({productId, productName, categoryName, description, price, 
         setIsAccordionOpen(!isAccordionOpen);
     };
 
-    function handleChange(e) {
+    function productChange(e) {
         const { name, value, files } = e.target;
         console.log(`Changing ${name} to ${value}`);
         if (name === 'category_name') {
@@ -106,37 +106,38 @@ const ProductItem = ({productId, productName, categoryName, description, price, 
 
     function handleDelete(e) {
         e.preventDefault();
+        if (window.confirm("Are you sure you want to delete this product?")) {
+            fetch(`http://localhost:8080/nexusHub/product/delete/${productId}`, {
+                method: "DELETE"
+            })
+                .then((response) => {
+                    if (response.ok) {
+                        console.log('Product deleted Successfully');
 
-        fetch(`http://localhost:8080/nexusHub/product/delete/${productId}`, {
-            method: "DELETE"
-        })
-            .then((response) => {
-                if (response.ok) {
-                    console.log('Product deleted Successfully');
-
-                    toast.success('Product deleted Successfully', {
-                        position: "top-center",
-                        draggable: true,
-                        hideProgressBar: true
-                    });
-
-                    setTimeout(() => {
-                        window.location.reload();
-                    }, 2000)
-
-                } else {
-                    console.error('Failed to delete product');
-                    toast.warn('Failed to delete product', {
+                        toast.success('Product deleted Successfully', {
                             position: "top-center",
                             draggable: true,
                             hideProgressBar: true
-                        }
-                    );
-                }
-            })
-            .catch((error) => {
-                console.error('Error:', error);
-            });
+                        });
+
+                        setTimeout(() => {
+                            window.location.reload();
+                        }, 2000)
+
+                    } else {
+                        console.error('Failed to delete product');
+                        toast.warn('Failed to delete product', {
+                                position: "top-center",
+                                draggable: true,
+                                hideProgressBar: true
+                            }
+                        );
+                    }
+                })
+                .catch((error) => {
+                    console.error('Error:', error);
+                });
+        }
     }
 
     return (
@@ -160,14 +161,14 @@ const ProductItem = ({productId, productName, categoryName, description, price, 
                             <h4>PRODUCT NAME</h4>
                         </label>
                         <div className="input-group mb-3 input-box">
-                            <input type="text" className="form-control" name="productName"  value={formData.productName} onChange={handleChange} required={true}></input>
+                            <input type="text" className="form-control" name="productName"  value={formData.productName} onChange={productChange} required={true}></input>
                         </div>
                         <label>
                             <h4>PRODUCT CATEGORY</h4>
                         </label>
                         <div className="input-group mb-3 input-box">
                             <div className="btn-group">
-                                <select className="form-select category-select" name="categoryId" value={formData.categoryId} onChange={handleChange} required={true}>
+                                <select className="form-select category-select" name="categoryId" value={formData.categoryId} onChange={productChange} required={true}>
                                     <option value="">Select Category</option>
                                     {categories.map(category => (
                                         <option key={category.categoryId} value={category.categoryId}>
@@ -181,26 +182,26 @@ const ProductItem = ({productId, productName, categoryName, description, price, 
                             <h4 className="">PRODUCT DESCRIPTION</h4>
                         </label>
                         <div className="input-group mb-3 input-box">
-                            <input type="text" className="form-control" name="description" value={formData.description} onChange={handleChange} required={true}></input>
+                            <input type="text" className="form-control" name="description" value={formData.description} onChange={productChange} required={true}></input>
                         </div>
                         <label>
                             <h4 className="">AVAILABLE QUANTITY</h4>
                         </label>
                         <div className="input-group mb-3 input-box">
-                            <input type="text" className="form-control" name="stock_quantity" value={formData.stock_quantity} onChange={handleChange} required={true}></input>
+                            <input type="text" className="form-control" name="stock_quantity" value={formData.stock_quantity} onChange={productChange} required={true}></input>
                         </div>
                         <label>
                             <h4 className="">PRICE</h4>
                         </label>
                         <div className="input-group mb-3 input-box">
                             <span className="input-group-text price-icon">$</span>
-                            <input type="text" className="form-control" name="price" value={formData.price} onChange={handleChange} required={true}></input>
+                            <input type="text" className="form-control" name="price" value={formData.price} onChange={productChange} required={true}></input>
                         </div>
                         <label>
                             <h4>IMAGE</h4>
                         </label>
                         <div className="input-group mb-3 input-box">
-                            <input type="file" className="form-control" name="image" id="formFileSm" onChange={handleChange}></input>
+                            <input type="file" className="form-control" name="image" id="formFileSm" onChange={productChange}></input>
                         </div>
                         <button type="submit" className="btn btn-primary">Save Product</button>
                     </form>
